Improve saving activation files

+++ sshrd.sh
@@ -50,6 +50,14 @@ fi
 
 chmod +x "$oscheck"/*
 
+if [ "$1" = '--reset-ssh' ]; then
+    "$oscheck"/iproxy 2222 22 &>/dev/null &
+    "$oscheck"/sshpass -p alpine ssh root@127.0.0.1 -p2222 -o StrictHostKeyChecking=no "/usr/sbin/nvram oblit-inprogress=5 || true"
+    "$oscheck"/sshpass -p alpine ssh root@127.0.0.1 -p2222 -o StrictHostKeyChecking=no "/sbin/reboot" 
+    echo "[*] Device should now show a progress bar and erase all data"
+    exit
+fi
+
 if [ "$1" = 'clean' ]; then
     rm -rf sshramdisk work
     echo "[*] Removed the current created SSH ramdisk"
@@ -115,13 +123,15 @@ elif [ "$1" = '--backup-activation' ]; then
         sleep .1
     fi
     "$oscheck"/iproxy 2222 22 &>/dev/null &
+    serial_number=$("$oscheck"/sshpass -p alpine ssh root@127.0.0.1 -p2222 -o StrictHostKeyChecking=no "/usr/sbin/ioreg -l | grep IOPlatformSerialNumber | sed 's/.*IOPlatformSerialNumber\" = \"\(.*\)\"/\1/' | cut -d '\"' -f4")
+    mkdir -p ./Activation_Records/$serial_number
     "$oscheck"/sshpass -p alpine ssh root@127.0.0.1 -p2222 -o StrictHostKeyChecking=no "/usr/bin/mount_filesystems || true"
-    "$oscheck"/sshpass -p alpine scp -P2222 -o StrictHostKeyChecking=no root@127.0.0.1:/mnt2/containers/Data/System/*/Library/activation_records/activation_record.plist . || true
-    "$oscheck"/sshpass -p alpine scp -P2222 -o StrictHostKeyChecking=no root@127.0.0.1:/mnt2/wireless/Library/Preferences/com.apple.commcenter.device_specific_nobackup.plist . || true
-    "$oscheck"/sshpass -p alpine scp -P2222 -o StrictHostKeyChecking=no root@127.0.0.1:/mnt2/mobile/Library/FairPlay/iTunes_Control/iTunes/IC-Info.sisv . || true
-    if [ -e activation_record.plist ] && [ -e com.apple.commcenter.device_specific_nobackup.plist ] && [ -e IC-Info.sisv ]; then
-    echo "[*] Activation files saved to SSHRD_Script directory"
-    elif [ -e activation_record.plist ] && [ -e com.apple.commcenter.device_specific_nobackup.plist ] && [ ! -e IC-Info.sisv ]; then
+    "$oscheck"/sshpass -p alpine scp -P2222 -o StrictHostKeyChecking=no root@127.0.0.1:/mnt2/containers/Data/System/*/Library/activation_records/activation_record.plist ./Activation_Records/$serial_number || true
+    "$oscheck"/sshpass -p alpine scp -P2222 -o StrictHostKeyChecking=no root@127.0.0.1:/mnt2/wireless/Library/Preferences/com.apple.commcenter.device_specific_nobackup.plist ./Activation_Records/$serial_number || true
+    "$oscheck"/sshpass -p alpine scp -P2222 -o StrictHostKeyChecking=no root@127.0.0.1:/mnt2/mobile/Library/FairPlay/iTunes_Control/iTunes/IC-Info.sisv ./Activation_Records/$serial_number || true
+    if [ -e Activation_Records/$serial_number/activation_record.plist ] && [ -e Activation_Records/$serial_number/com.apple.commcenter.device_specific_nobackup.plist ] && [ -e Activation_Records/$serial_number/IC-Info.sisv ]; then
+    echo "[*] Activation files saved to Activation_Records/$serial_number"
+    elif [ -e Activation_Records/$serial_number/activation_record.plist ] && [ -e Activation_Records/$serial_number/com.apple.commcenter.device_specific_nobackup.plist ] && [ ! -e Activation_Records/$serial_number/IC-Info.sisv ]; then
     echo "[*] Failed to save IC-Info.sisv, delete current /mnt2/mobile/Library/FairPlay/iTunes_Control/iTunes/IC-Info.sisv, reboot to lock screen, enter DFU mode, boot SSH ramdisk and try again"
     else
     echo "[*] Failed to save activation files, select a ramdisk version that is identical or close enough to device's version and try again"
@@ -130,10 +140,6 @@ elif [ "$1" = '--backup-activation' ]; then
     sudo killall usbmuxd 2>/dev/null | true
     exit
 elif [ "$1" = '--restore-activation' ]; then
-    if [ ! -e activation_record.plist ]; then
-        echo "[*] Activation files not found"
-        exit
-    fi
     if [ "$oscheck" = 'Linux' ]; then
         sudo systemctl stop usbmuxd 2>/dev/null | true
         sudo killall usbmuxd 2>/dev/null | true
@@ -142,12 +148,17 @@ elif [ "$1" = '--restore-activation' ]; then
         sleep .1
     fi
     "$oscheck"/iproxy 2222 22 &>/dev/null &
+    serial_number=$("$oscheck"/sshpass -p alpine ssh root@127.0.0.1 -p2222 -o StrictHostKeyChecking=no "/usr/sbin/ioreg -l | grep IOPlatformSerialNumber | sed 's/.*IOPlatformSerialNumber\" = \"\(.*\)\"/\1/' | cut -d '\"' -f4")
+    if [ ! -e Activation_Records/$serial_number/activation_record.plist ]; then
+        echo "[*] Activation files not found"
+        exit
+    fi
     "$oscheck"/sshpass -p alpine ssh root@127.0.0.1 -p2222 -o StrictHostKeyChecking=no "/usr/bin/mount_filesystems || true"
     "$oscheck"/sshpass -p alpine ssh root@127.0.0.1 -p2222 -o StrictHostKeyChecking=no "rm -rf /mnt2/mobile/Media/Downloads/Activation /mnt2/mobile/Media/Activation"
     "$oscheck"/sshpass -p alpine ssh root@127.0.0.1 -p2222 -o StrictHostKeyChecking=no "mkdir -p /mnt2/mobile/Media/Downloads/Activation"
-    "$oscheck"/sshpass -p alpine scp -P2222 -o StrictHostKeyChecking=no activation_record.plist root@127.0.0.1:/mnt2/mobile/Media/Downloads/Activation
-    "$oscheck"/sshpass -p alpine scp -P2222 -o StrictHostKeyChecking=no com.apple.commcenter.device_specific_nobackup.plist root@127.0.0.1:/mnt2/mobile/Media/Downloads/Activation
-    "$oscheck"/sshpass -p alpine scp -P2222 -o StrictHostKeyChecking=no IC-Info.sisv root@127.0.0.1:/mnt2/mobile/Media/Downloads/Activation
+    "$oscheck"/sshpass -p alpine scp -P2222 -o StrictHostKeyChecking=no Activation_Records/$serial_number/activation_record.plist root@127.0.0.1:/mnt2/mobile/Media/Downloads/Activation
+    "$oscheck"/sshpass -p alpine scp -P2222 -o StrictHostKeyChecking=no Activation_Records/$serial_number/com.apple.commcenter.device_specific_nobackup.plist root@127.0.0.1:/mnt2/mobile/Media/Downloads/Activation
+    "$oscheck"/sshpass -p alpine scp -P2222 -o StrictHostKeyChecking=no Activation_Records/$serial_number/IC-Info.sisv root@127.0.0.1:/mnt2/mobile/Media/Downloads/Activation
     "$oscheck"/sshpass -p alpine ssh root@127.0.0.1 -p2222 -o StrictHostKeyChecking=no "mv -f /mnt2/mobile/Media/Downloads/Activation /mnt2/mobile/Media"
     "$oscheck"/sshpass -p alpine ssh root@127.0.0.1 -p2222 -o StrictHostKeyChecking=no "/usr/sbin/chown -R mobile:mobile /mnt2/mobile/Media/Activation"
     "$oscheck"/sshpass -p alpine ssh root@127.0.0.1 -p2222 -o StrictHostKeyChecking=no "chmod -R 755 /mnt2/mobile/Media/Activation"
@@ -176,15 +187,17 @@ elif [ "$1" = '--backup-activation-hfs' ]; then
         sleep .1
     fi
     "$oscheck"/iproxy 2222 22 &>/dev/null &
+    serial_number=$("$oscheck"/sshpass -p alpine ssh root@127.0.0.1 -p2222 -o StrictHostKeyChecking=no "/usr/sbin/ioreg -l | grep IOPlatformSerialNumber | sed 's/.*IOPlatformSerialNumber\" = \"\(.*\)\"/\1/' | cut -d '\"' -f4")
+    mkdir -p ./Activation_Records/$serial_number
     "$oscheck"/sshpass -p alpine ssh root@127.0.0.1 -p2222 -o StrictHostKeyChecking=no "/sbin/mount_hfs /dev/disk0s1s1 /mnt1 || true"
     "$oscheck"/sshpass -p alpine ssh root@127.0.0.1 -p2222 -o StrictHostKeyChecking=no "/usr/libexec/seputil --load /mnt1/usr/standalone/firmware/sep-firmware.img4 || true"
     "$oscheck"/sshpass -p alpine ssh root@127.0.0.1 -p2222 -o StrictHostKeyChecking=no "/sbin/mount_hfs /dev/disk0s1s2 /mnt2 || true"
-    "$oscheck"/sshpass -p alpine scp -P2222 -o StrictHostKeyChecking=no root@127.0.0.1:/mnt2/containers/Data/System/*/Library/activation_records/activation_record.plist . || true
-    "$oscheck"/sshpass -p alpine scp -P2222 -o StrictHostKeyChecking=no root@127.0.0.1:/mnt2/wireless/Library/Preferences/com.apple.commcenter.device_specific_nobackup.plist . || true
-    "$oscheck"/sshpass -p alpine scp -P2222 -o StrictHostKeyChecking=no root@127.0.0.1:/mnt2/mobile/Library/FairPlay/iTunes_Control/iTunes/IC-Info.sisv . || true
-    if [ -e activation_record.plist ] && [ -e com.apple.commcenter.device_specific_nobackup.plist ] && [ -e IC-Info.sisv ]; then
-    echo "[*] Activation files saved to SSHRD_Script directory"
-    elif [ -e activation_record.plist ] && [ -e com.apple.commcenter.device_specific_nobackup.plist ] && [ ! -e IC-Info.sisv ]; then
+    "$oscheck"/sshpass -p alpine scp -P2222 -o StrictHostKeyChecking=no root@127.0.0.1:/mnt2/containers/Data/System/*/Library/activation_records/activation_record.plist ./Activation_Records/$serial_number || true
+    "$oscheck"/sshpass -p alpine scp -P2222 -o StrictHostKeyChecking=no root@127.0.0.1:/mnt2/wireless/Library/Preferences/com.apple.commcenter.device_specific_nobackup.plist ./Activation_Records/$serial_number || true
+    "$oscheck"/sshpass -p alpine scp -P2222 -o StrictHostKeyChecking=no root@127.0.0.1:/mnt2/mobile/Library/FairPlay/iTunes_Control/iTunes/IC-Info.sisv ./Activation_Records/$serial_number || true
+    if [ -e Activation_Records/$serial_number/activation_record.plist ] && [ -e Activation_Records/$serial_number/com.apple.commcenter.device_specific_nobackup.plist ] && [ -e Activation_Records/$serial_number/IC-Info.sisv ]; then
+    echo "[*] Activation files saved to Activation_Records/$serial_number"
+    elif [ -e Activation_Records/$serial_number/activation_record.plist ] && [ -e Activation_Records/$serial_number/com.apple.commcenter.device_specific_nobackup.plist ] && [ ! -e Activation_Records/$serial_number/IC-Info.sisv ]; then
     echo "[*] Failed to save IC-Info.sisv, delete current /mnt2/mobile/Library/FairPlay/iTunes_Control/iTunes/IC-Info.sisv, reboot to lock screen, enter DFU mode, boot SSH ramdisk and try again"
     else
     echo "[*] Failed to save activation files, select a ramdisk version that is identical or close enough to device's version and try again"
@@ -193,10 +206,6 @@ elif [ "$1" = '--backup-activation-hfs' ]; then
     sudo killall usbmuxd 2>/dev/null | true
     exit
 elif [ "$1" = '--restore-activation-hfs' ]; then
-    if [ ! -e activation_record.plist ]; then
-        echo "[*] Activation files not found"
-        exit
-    fi
     if [ "$oscheck" = 'Linux' ]; then
         sudo systemctl stop usbmuxd 2>/dev/null | true
         sudo killall usbmuxd 2>/dev/null | true
@@ -205,15 +214,21 @@ elif [ "$1" = '--restore-activation-hfs' ]; then
         sleep .1
     fi
     "$oscheck"/iproxy 2222 22 &>/dev/null &
+    serial_number=$("$oscheck"/sshpass -p alpine ssh root@127.0.0.1 -p2222 -o StrictHostKeyChecking=no "/usr/sbin/ioreg -l | grep IOPlatformSerialNumber | sed 's/.*IOPlatformSerialNumber\" = \"\(.*\)\"/\1/' | cut -d '\"' -f4")
+    if [ ! -e Activation_Records/$serial_number/activation_record.plist ]; then
+        echo "[*] Activation files not found"
+        exit
+    fi
+    "$oscheck"/iproxy 2222 22 &>/dev/null &
     "$oscheck"/sshpass -p alpine ssh root@127.0.0.1 -p2222 -o StrictHostKeyChecking=no "/sbin/mount_hfs /dev/disk0s1s1 /mnt1 || true"
     "$oscheck"/sshpass -p alpine ssh root@127.0.0.1 -p2222 -o StrictHostKeyChecking=no "/usr/libexec/seputil --load /mnt1/usr/standalone/firmware/sep-firmware.img4 || true"
     "$oscheck"/sshpass -p alpine ssh root@127.0.0.1 -p2222 -o StrictHostKeyChecking=no "/sbin/mount_hfs /dev/disk0s1s2 /mnt2 || true"
     "$oscheck"/sshpass -p alpine ssh root@127.0.0.1 -p2222 -o StrictHostKeyChecking=no "rm -rf /mnt2/mobile/Media/Downloads/Activation /mnt2/mobile/Media/Activation"
     "$oscheck"/sshpass -p alpine ssh root@127.0.0.1 -p2222 -o StrictHostKeyChecking=no "mkdir -p /mnt2/mobile/Media/Downloads/Activation"
-    "$oscheck"/sshpass -p alpine scp -P2222 -o StrictHostKeyChecking=no activation_record.plist root@127.0.0.1:/mnt2/mobile/Media/Downloads/Activation
-    "$oscheck"/sshpass -p alpine scp -P2222 -o StrictHostKeyChecking=no com.apple.commcenter.device_specific_nobackup.plist root@127.0.0.1:/mnt2/mobile/Media/Downloads/Activation
-    "$oscheck"/sshpass -p alpine scp -P2222 -o StrictHostKeyChecking=no IC-Info.sisv root@127.0.0.1:/mnt2/mobile/Media/Downloads/Activation
-    "$oscheck"/sshpass -p alpine ssh root@127.0.0.1 -p2222 -o StrictHostKeyChecking=no "mv -f /mnt2/mobile/Media/Downloads/Activation /mnt2/mobile/Media"
+    "$oscheck"/sshpass -p alpine scp -P2222 -o StrictHostKeyChecking=no Activation_Records/$serial_number/activation_record.plist root@127.0.0.1:/mnt2/mobile/Media/Downloads/Activation
+    "$oscheck"/sshpass -p alpine scp -P2222 -o StrictHostKeyChecking=no Activation_Records/$serial_number/com.apple.commcenter.device_specific_nobackup.plist root@127.0.0.1:/mnt2/mobile/Media/Downloads/Activation
+    "$oscheck"/sshpass -p alpine scp -P2222 -o StrictHostKeyChecking=no Activation_Records/$serial_number/IC-Info.sisv root@127.0.0.1:/mnt2/mobile/Media/Downloads/Activation
+   "$oscheck"/sshpass -p alpine ssh root@127.0.0.1 -p2222 -o StrictHostKeyChecking=no "mv -f /mnt2/mobile/Media/Downloads/Activation /mnt2/mobile/Media"
     "$oscheck"/sshpass -p alpine ssh root@127.0.0.1 -p2222 -o StrictHostKeyChecking=no "/usr/sbin/chown -R mobile:mobile /mnt2/mobile/Media/Activation"
     "$oscheck"/sshpass -p alpine ssh root@127.0.0.1 -p2222 -o StrictHostKeyChecking=no "chmod -R 755 /mnt2/mobile/Media/Activation"
     "$oscheck"/sshpass -p alpine ssh root@127.0.0.1 -p2222 -o StrictHostKeyChecking=no "cd /mnt2/containers/Data/System/*/Library/internal; mkdir -p ../activation_records"
